Feat : 마이페이지

+++ src/main/java/hongik/map/honggildong/domain/member/controller/MemberController.java
@@ -11,7 +11,10 @@ import hongik.map.honggildong.domain.review.dto.ReviewResponseDTO;
 import hongik.map.honggildong.domain.review.entity.Review;
 import hongik.map.honggildong.domain.review.service.ReviewService;
 import hongik.map.honggildong.global.apiPayload.ApiResponse;
+import hongik.map.honggildong.global.apiPayload.code.status.ErrorStatus;
+import hongik.map.honggildong.global.apiPayload.exception.GeneralException;
 import hongik.map.honggildong.global.security.service.CustomUserDetails;
+import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springdoc.core.annotations.ParameterObject;
@@ -33,8 +36,14 @@ public class MemberController {
     private final LikeService likeService;
 
     //회원 정보 조회
-    @GetMapping("")
+    @GetMapping("/profile")
+    @Operation(summary = "회원 정보 조회")
     public ApiResponse<MemberResponseDTO.General> getMember(@AuthenticationPrincipal CustomUserDetails userDetails) {
+
+        if(userDetails==null){
+            throw new GeneralException(ErrorStatus.UNAUTHORIZED);
+        }
+
         Member member = memberService.getMemberByUserDetails(userDetails);
         MemberResponseDTO.General body = MemberConverter.toGeneralDTO(member);
 
@@ -43,8 +52,13 @@ public class MemberController {
 
     //회원 프로필 변경
     @PatchMapping("/mypage/profile")
+    @Operation(summary = "회원 프로필 수정", description = "프로필 사진 삭제 시 빈 문자열로 보낼 것")
     public ApiResponse<MemberResponseDTO.General> updateMember(@AuthenticationPrincipal CustomUserDetails userDetails,
                                                             @RequestBody MemberRequestDTO.UpdateProfile request) {
+
+        if(userDetails==null){
+            throw new GeneralException(ErrorStatus.UNAUTHORIZED);
+        }
         Member member = memberService.getMemberByUserDetails(userDetails);
         MemberResponseDTO.General body = MemberConverter.toGeneralDTO(memberService.updateProfile(member, request));
 
@@ -54,6 +68,11 @@ public class MemberController {
     //회원 탈퇴
     @DeleteMapping("/delete")
     public ApiResponse<String> deleteMember(@AuthenticationPrincipal CustomUserDetails userDetails) {
+
+        if(userDetails==null){
+            throw new GeneralException(ErrorStatus.UNAUTHORIZED);
+        }
+
         Member member = memberService.getMemberByUserDetails(userDetails);
         memberService.deleteMember(member);
 
@@ -62,23 +81,30 @@ public class MemberController {
 
     //내가 쓴 리뷰 리스트 조회
     @GetMapping("/mypage/reviews")
+    @Operation(summary = "내가 쓴 리뷰 조회")
     public ApiResponse<ReviewResponseDTO.MyGeneralPage> getMyReviews(@AuthenticationPrincipal CustomUserDetails userDetails,
                                                                      @ParameterObject Pageable pageable) {
-        Member member = memberService.getMemberByUserDetails(userDetails);
-        Page<Review> reviews = reviewService.getReviewListOf(member,pageable);
-        List<Long> likedReviewIds = new ArrayList<>();
-        ReviewResponseDTO.MyGeneralPage body = ReviewConverter.toMyGeneralPage(reviews,likedReviewIds);
+        if(userDetails==null){
+            throw new GeneralException(ErrorStatus.UNAUTHORIZED);
+        }
+
+        Member member = userDetails.getMember();
+
+        ReviewResponseDTO.MyGeneralPage body = reviewService.getReviewListOf(member,pageable);
 
         return ApiResponse.onSuccess(body);
     }
 
     //내가 좋아요한 리뷰 리스트 조회
     @GetMapping("/mypage/likes")
-    public ApiResponse<Page<ReviewResponseDTO.General>> getMyLikes(@AuthenticationPrincipal CustomUserDetails userDetails,
-                                                                   @ParameterObject Pageable pageable) {
-        Member member = memberService.getMemberByUserDetails(userDetails);
-        Page<Review> reviews = likeService.getLikedReviewListOf(member,pageable);
-        Page<ReviewResponseDTO.General> body = reviews.map(rw->ReviewConverter.toGeneralDTO(rw,true));
+    public ApiResponse<ReviewResponseDTO.GeneralPage> getMyLikes(@AuthenticationPrincipal CustomUserDetails userDetails,
+                                                                 @ParameterObject Pageable pageable) {
+        if(userDetails==null){
+            throw new GeneralException(ErrorStatus.UNAUTHORIZED);
+        }
+
+        Member member = userDetails.getMember();
+        ReviewResponseDTO.GeneralPage body = likeService.getLikedReviewListOf(member,pageable);
 
         return ApiResponse.onSuccess(body);
     }
